Update punishments system

+++ multibot/models/punishments.py
@@ -3,9 +3,8 @@ from __future__ import annotations  # todo0 remove in 3.11
 __all__ = ['PunishmentBase', 'Ban', 'Mute']
 
 import datetime
-import itertools
-from dataclasses import dataclass
-from typing import Any, Callable, Iterator
+from dataclasses import dataclass, field
+from typing import Any, Callable
 
 import flanautils
 from flanautils import DCMongoBase, FlanaBase
@@ -19,8 +18,7 @@ from multibot.models.message import Message
 
 @dataclass(eq=False)
 class PunishmentBase(DCMongoBase, FlanaBase):
-    _unique_keys = ('platform', 'user_id', 'group_id', 'until', 'is_active')
-    _nullable_unique_keys = ('until',)
+    _unique_keys = ('platform', 'user_id', 'group_id')
 
     platform: Platform = None
     user_id: int = None
@@ -28,6 +26,7 @@ class PunishmentBase(DCMongoBase, FlanaBase):
     time: int | datetime.timedelta = None
     until: datetime.datetime = None
     is_active: bool = True
+    last_update: datetime.datetime = field(default_factory=lambda: datetime.datetime.now(datetime.timezone.utc))
 
     def __post_init__(self):
         super().__post_init__()
@@ -37,38 +36,17 @@ class PunishmentBase(DCMongoBase, FlanaBase):
         if self.time:
             self.until = datetime.datetime.now(datetime.timezone.utc) + self.time
 
-    @classmethod
-    def _get_grouped_punishments(cls, platform: Platform) -> tuple[tuple[tuple[int, int], list[PunishmentBase]], ...]:
-        sorted_punishments = cls.find({'platform': platform.value}, sort_keys=('user_id', 'group_id', 'until'))
-        group_iterator: Iterator[
-            tuple[
-                tuple[int, int],
-                Iterator[cls]
-            ]
-        ] = itertools.groupby(sorted_punishments, key=lambda punishment: (punishment.user_id, punishment.group_id))
-        return tuple(((user_id, group_id), list(group_)) for (user_id, group_id), group_ in group_iterator)
-
     def _mongo_repr(self) -> Any:
         return {
             'platform': self.platform.value,
             'user_id': self.user_id,
             'group_id': self.group_id,
             'until': self.until,
-            'is_active': self.is_active
+            'is_active': self.is_active,
+            'last_update': self.last_update
         }
 
-    @classmethod
-    async def check_olds(cls, unpunishment_method: Callable, platform: Platform):
-        punishment_groups = cls._get_grouped_punishments(platform)
-
-        now = datetime.datetime.now(datetime.timezone.utc)
-        for (_, _), sorted_punishments in punishment_groups:
-            if (last_punishment := sorted_punishments[-1]).until and last_punishment.until <= now:
-                await last_punishment.unpunish(unpunishment_method)
-                for punishment in sorted_punishments:
-                    punishment.delete()
-
-    async def punish(self, punishment_method: Callable, unpunishment_method: Callable, message: Message = None):
+    async def apply(self, punishment_method: Callable, unpunishment_method: Callable, message: Message = None):
         try:
             await punishment_method(self.user_id, self.group_id, message)
         except (BadRoleError, UserDisconnectedError) as e:
@@ -77,11 +55,19 @@ class PunishmentBase(DCMongoBase, FlanaBase):
             else:
                 raise e
         else:
-            self.save()
+            self.save(pull_exclude_fields=('until',))
             if self.time is not None and datetime.timedelta() <= self.time <= constants.TIME_THRESHOLD_TO_MANUAL_UNPUNISH:
                 await flanautils.do_later(self.time, self.check_olds, unpunishment_method, self.platform)
 
-    async def unpunish(self, unpunishment_method: Callable, message: Message = None):
+    @classmethod
+    async def check_olds(cls, unpunishment_method: Callable, platform: Platform):
+        punishments = cls.find({'platform': platform.value})
+
+        for punishment in punishments:
+            if punishment.until and punishment.until <= datetime.datetime.now(datetime.timezone.utc):
+                await punishment.remove(unpunishment_method)
+
+    async def remove(self, unpunishment_method: Callable, message: Message = None, delete=True):
         try:
             await unpunishment_method(self.user_id, self.group_id, message)
         except UserDisconnectedError as e:
@@ -90,15 +76,15 @@ class PunishmentBase(DCMongoBase, FlanaBase):
             else:
                 raise e
         else:
-            try:
-                self.__class__.find_one({
-                    'platform': self.platform.value,
-                    'user_id': self.user_id,
-                    'group_id': self.group_id,
-                    'until': None
-                }).delete()
-            except AttributeError:
-                pass
+            if delete:
+                try:
+                    self.__class__.find_one({
+                        'platform': self.platform.value,
+                        'user_id': self.user_id,
+                        'group_id': self.group_id
+                    }).delete()
+                except AttributeError:
+                    pass
 
 
 @dataclass(eq=False)
